Bump go version in go.mod to 1.19
Current version in go.mod is 1.13 but in practice you need 1.19 to compile the project
so the version in go.mod should be changed to reflect this.

There were discussions about bumping this to a version that we test in CI but
users that aren't using the latest go version would be significantly impacted by this
so we decided against it.

Patch by João Reis; reviewed by Stanislav Bychkov, James Hartig for CASSGO-34